+++ tests/mocks/mockInstabugUtils.js
@@ -1,0 +1,11 @@
+jest.mock('../../src/utils/InstabugUtils', () => {
+  return {
+    parseErrorStack: jest.fn(),
+    captureJsErrors: jest.fn(),
+    setOnReportHandler: jest.fn(),
+    isOnReportHandlerSet: jest.fn(),
+    getActiveRouteName: jest.fn(),
+    stringifyIfNotString: jest.fn(),
+    getStackTrace: jest.fn().mockReturnValue('javascriptStackTrace'),
+  };
+});